+++ createElem.js
@@ -1,0 +1,9 @@
+function createElement(tag, className) {
+	const $tag = document.createElement(tag);
+	if(className) {
+		$tag.classList.add(className);
+	}
+	return $tag;
+};
+
+export default createElement;
