+++ globalStor.js
@@ -1,0 +1,13 @@
+export const GlobalVar = {
+	$arenas: document.querySelector('.root .arenas'),
+	$formFight: document.querySelector('.control'),
+	$chat: document.querySelector('.chat'),
+	$fightButton: document.querySelector('.button'),
+	HIT: {
+		head: 30,
+		body: 25,
+		foot: 20,
+	},
+	PENALTY: 50,
+	ATTACK: ['head', 'body', 'foot'],
+};
